backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
boto3==1.42.1
botocore==1.42.1
certifi==2025.11.12
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.6.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
s3transfer==0.16.0
sentry-sdk==2.45.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1